Question 7
Using the sourcetyoe "aws:s3:accesslogs" and the known hostname "splunk.froth.ly" I searched for anything using "Put" as that would be used in a PutObject operation. As we know the uploaded file is a text file, i searched for listings containing "docx" firstly but no results appeared, I then tried "txt" and got a match, to confirm, i made a note of the timestamp and compared it to the time we know the bucket was open for. The bucket was mad public at 2:01:46.000 PM and the file was uploaded at 2:02:44.000 PM meaning after the bucket was made public.

Answer: OPEN_BUCKET_PLEASE_FIX.txt